add extra p2p functions to pkg
