+++ cmd/tendermint/commands/show_node_id.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/tendermint/tendermint/internal/p2p"
+	"github.com/tendermint/tendermint/pkg/p2p"
 )
 
 // ShowNodeIDCmd dumps node's ID to the standard output.
